WCM-34 Improve dependency handling
Web Content and Document and Media referenced by rules, tracking actions or display portlets will be exported/imported. Any other referenced content will be exported as a reference and imported if available in the lar file. Other entities will be exported as references and should be available in the target environment beforehand.